Move integration tests to packages/integration-tests
