+++ packages/integration-tests/src/elicitation.test.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import { describe, it, expect, afterEach } from "vitest";
-import { type UserConfig } from "../../src/common/config/userConfig.js";
+import { type UserConfig } from "mongodb-mcp-server";
 import { defaultTestConfig } from "@mongodb-js/mcp-test-utils";
-import { Elicitation } from "../../src/elicitation.js";
+import { Elicitation } from "mongodb-mcp-server";
 import { createMockElicitInput } from "@mongodb-js/mcp-test-utils";
 import { describeWithMongoDB } from "@mongodb-js/mcp-test-utils";
