+++ packages/integration-tests/src/resources/exportedData.test.ts
@@ -6,7 +6,7 @@ import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
 import { defaultTestConfig, getDataFromUntrustedContent, resourceChangedNotification, timeout } from "@mongodb-js/mcp-test-utils";
 import { describeWithMongoDB } from "@mongodb-js/mcp-test-utils";
 import { contentWithResourceURILink } from "../tools/mongodb/read/export.test.js";
-import type { UserConfig } from "../../../src/lib.js";
+import type { UserConfig } from "mongodb-mcp-server";
 
 const userConfig: UserConfig = {
     ...defaultTestConfig,
