+++ packages/integration-tests/src/setup/agentIdMapping.test.ts
@@ -3,8 +3,8 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { afterAll, beforeAll, describe, it, expect } from "vitest";
-import { AI_TOOL_REGISTRY, type AIToolType } from "../../../src/setup/aiTool.js";
-import { buildSkillsAddArgs } from "../../../src/setup/installSkills.js";
+import { AI_TOOL_REGISTRY, type AIToolType } from "mongodb-mcp-server/setup";
+import { buildSkillsAddArgs } from "mongodb-mcp-server/setup";
 
 /**
  * Verifies every `AIToolType`'s `skillsAgentId` returns a value the pinned
