+++ src/setup/index.ts
@@ -1,0 +1,2 @@
+export { AI_TOOL_REGISTRY, type AIToolType } from "./aiTool.js";
+export { buildSkillsAddArgs } from "./installSkills.js";
