+++ vitest.config.ts
@@ -10,7 +10,7 @@ const vitestDefaultExcludes = [
     "**/{karma,rollup,webpack,vite,vitest,jest,ava,babel,nyc,cypress,tsup,build,eslint,prettier}.config.*",
 ];
 
-const longRunningTests = ["tests/integration/tools/atlas/performanceAdvisor.test.ts"];
+const longRunningTests = ["packages/integration-tests/src/tools/atlas/performanceAdvisor.test.ts"];
 
 if (process.env.SKIP_ATLAS_INTEGRATION_TESTS === "true") {
     vitestDefaultExcludes.push("**/integration/**/atlas/**");
@@ -26,7 +26,7 @@ export default defineConfig({
         pool: "threads",
         testTimeout: 3600000,
         hookTimeout: 3600000,
-        setupFiles: ["./tests/setup.ts"],
+        setupFiles: ["./packages/integration-tests/src/setup.ts"],
         coverage: {
             exclude: [
                 // Required: import.meta.glob() in src/ui creates Vite virtual modules (\0 prefixed paths)
